+++ app/code/ThirdParty/BlogArticle/Block/Adminhtml/Post/Edit.php
@@ -83,6 +83,7 @@ class Edit extends Template
     {
         $collection = $this->categoryCollectionFactory->create();
         $collection->addFieldToFilter('is_active', 1);
+        $collection->setOrder('sort_order', 'ASC');
         $collection->setOrder('name', 'ASC');
         return $collection;
     }