+++ app/code/ThirdParty/BlogArticle/Model/Resolver/Categories.php
@@ -36,6 +36,7 @@ class Categories implements ResolverInterface
                 'url_key' => $category->getUrlKey(),
                 'description' => $category->getDescription(),
                 'is_active' => $category->getIsActive(),
+                'sort_order' => $category->getSortOrder(),
                 'creation_time' => $category->getCreationTime(),
                 'update_time' => $category->getUpdateTime(),
             ];