+++ app/code/ThirdParty/BlogArticle/Model/Sitemap/ItemProvider/Categories.php
@@ -25,6 +25,7 @@ class Categories implements ItemProviderInterface
     {
         $collection = $this->collectionFactory->create();
         $collection->addFieldToFilter('is_active', 1);
+        $collection->setOrder('sort_order', 'ASC');
         $collection->setOrder('name', 'ASC');
 
         $items = [];